requirements.txt
django
pathlib
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-simpleui
django-extensions
pymysql
django-filter